backend/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-filter==21.1
djoser==2.1.0
drf-extra-fields==3.2.1
pillow==8.4.0
gunicorn==20.1.0
flake8==4.0.1
psycopg2-binary
drf-yasg==1.20.0